[server-actions]: Now we are fetching data for country dropdown from a server action function instead of a fetch request.
