+++ app/components/CountryAutocomplete.tsx
@@ -1,14 +1,14 @@
 "use client";
 import {
   Autocomplete,
-  FilterOptionsState,
   TextField,
   createFilterOptions,
 } from "@mui/material";
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { Person } from "../lib/person";
 import { Country } from "../lib/country";
 import CircularProgress from '@mui/material/CircularProgress';
+import { fetchCountries } from "../lib/countryActions";
 
 /*
 DONE: 1. When the user does not select anything, nothing should show in the result list. DONE 
@@ -16,8 +16,9 @@ DONE: 2. The result list only shows when the user types in at least 1 character.
 DONE: 3. the result should be filtered based on "startWith" logic. eg when I type Melb, then, all others that doesn't startWith Melb should not show up DONE
 DONE: 4. there's an API search each time (or via Server Action) each time the user enters data. The filtering should happen on server side, not on the client side. DONE
 DONE: 5. Add throttling in the server requests via country auto-complete.
-TODO: 6. Add a loader in the country auto-complete.
+DONE: 6. Add a loader in the country auto-complete.
 TODO: 7. Try to remove countryId field while adding or editing person. As we are already sending it in country key.
+DONE: 8: Convert the fetchCountries function in the useEffect() to server action function.
 */
 
 interface CountryAutocompleteProps {
@@ -36,38 +37,20 @@ const CountryAutocomplete: React.FC<CountryAutocompleteProps> = ({
   const [loading, setLoading] = useState(false);
 
   useEffect(() => {
-    const fetchCountries = async () => {
+    const fetchCountriesData = async () => {
       if (inputValue.trim().length === 0) {
         setOptions([]);
         return;
       }
       setLoading(true);
-      try {
-        const response = await fetch(
-          "/api/country?" + new URLSearchParams({ searchTerm: inputValue })
-        );
-        if (response.ok) {
-          const data = await response.json();
-          const isExisting = data.some(
-            (option: Country) => inputValue === option.name
-          );
-          if (inputValue !== "" && !isExisting) {
-            data.push({
-              inputValue,
-              name: `Add "${inputValue}"`,
-            });
-          }
-          setOptions(data);
-        } else {
-          console.log("Error fetching country data");
-        }
-      } catch (error) {
-        console.log("Error fetching country data:", error);
-      }
+      
+      const data = await fetchCountries(inputValue);
+      setOptions(data);
+
       setLoading(false);
     };
 
-    fetchCountries();
+    fetchCountriesData();
   }, [inputValue]);
 
   let timeoutTimer:ReturnType<typeof setTimeout> | null;
